Make RunnableParameterizable as a common base class for directors, acquisitions and experiments (stuff with a run()-function).
