+++ concert/experiments/base.py
@@ -15,8 +15,9 @@ import concert.devices.base
 from concert.coroutines.base import background, broadcast, start
 from concert.coroutines.sinks import count
 from concert.progressbar import wrap_iterable
-from concert.base import check, Parameterizable, Parameter, Selection, State, StateError
-from concert.helpers import get_state_from_awaitable, get_basename
+from concert.base import (check, Parameter, Selection, StateError,
+                          RunnableParameterizable)
+from concert.helpers import get_basename
 from concert.loghandler import AsyncLoggingHandlerCloser
 from functools import partial
 
@@ -55,7 +56,7 @@ class Consumer:
         LOG.debug('%s finished in %.3f s', self._corofunc.__qualname__, time.perf_counter() - st)
 
 
-class Acquisition(Parameterizable):
+class Acquisition(RunnableParameterizable):
     """
     An acquisition acquires data, gets it and sends it to consumers. This is a base class for local
     and remote acquisitions and must not be used directly.
@@ -77,7 +78,6 @@ class Acquisition(Parameterizable):
         a coroutine function which acquires the data, takes no arguments, can be None.
 
     """
-    state = State(default='standby')
 
     async def __ainit__(self, name, producer_corofunc, producer=None, acquire=None):
         self.name = name
@@ -97,11 +97,7 @@ class Acquisition(Parameterizable):
         if acquire and not asyncio.iscoroutinefunction(acquire):
             raise TypeError('acquire must be a coroutine function')
         self.acquire = acquire
-        self._run_awaitable = None
-        await Parameterizable.__ainit__(self)
-
-    async def _get_state(self):
-        return await get_state_from_awaitable(self._run_awaitable)
+        await super().__ainit__()
 
     @background
     async def _run(self):
@@ -226,7 +222,7 @@ class Acquisition(Parameterizable):
         return "Acquisition({})".format(self.name)
 
 
-class Experiment(Parameterizable):
+class Experiment(RunnableParameterizable):
 
     """
     Experiment base class. An experiment can be run multiple times with the output data and log
@@ -265,7 +261,6 @@ class Experiment(Parameterizable):
     separate_scans = Parameter()
     name_fmt = Parameter()
     current_name = Parameter(help="Name of the current iteration")
-    state = State(default='standby')
     log_level = Selection(['critical', 'error', 'warning', 'info', 'debug'])
     log_devices_at_start = Parameter()
     log_devices_at_finish = Parameter()
@@ -286,8 +281,7 @@ class Experiment(Parameterizable):
         self._log_devices_at_start = None
         self._log_devices_at_finish = None
         self.ready_to_prepare_next_sample = asyncio.Event()
-        self._run_awaitable = None
-        await Parameterizable.__ainit__(self)
+        await super().__ainit__()
         await self.set_log_devices_at_start(True)
         await self.set_log_devices_at_finish(True)
 
@@ -315,7 +309,8 @@ class Experiment(Parameterizable):
 
         :param name: Name of the device
         :param device: Device to log
-        :param optional: If True, an exception when trying to log the device will not cause an error.
+        :param optional: If True, an exception when trying to log the device will not cause an
+            error.
         """
         if optional:
             self._devices_to_log_optional[name] = device
@@ -462,15 +457,6 @@ class Experiment(Parameterizable):
             await acq()
 
     @background
-    @check(source=['standby', 'error', 'cancelled'], target=['standby', 'cancelled'])
-    async def run(self):
-        self._run_awaitable = self._run()
-        await self._run_awaitable
-
-    async def _get_state(self):
-        return await get_state_from_awaitable(self._run_awaitable)
-
-    @background
     async def _run(self):
         self.ready_to_prepare_next_sample.clear()
         start_time = time.time()
@@ -490,7 +476,8 @@ class Experiment(Parameterizable):
                 self.log.addHandler(handler)
                 if await self.get_log_devices_at_start():
                     exp_metadata: str = await self._prepare_metadata_str()
-                    await self.walker.log_to_json(payload=exp_metadata, filename="experiment_start.json")
+                    await self.walker.log_to_json(payload=exp_metadata,
+                                                  filename="experiment_start.json")
             self._current_name = get_basename(await self.walker.get_current())
         self.log.info(await self.info_table)
         for name, device in self._devices_to_log.items():
@@ -507,7 +494,8 @@ class Experiment(Parameterizable):
                 if self.walker:
                     if await self.get_log_devices_at_finish():
                         exp_metadata: str = await self._prepare_metadata_str()
-                        await self.walker.log_to_json(payload=exp_metadata, filename="experiment_finish.json")
+                        await self.walker.log_to_json(payload=exp_metadata,
+                                                      filename="experiment_finish.json")
             except Exception as e:
                 LOG.warning(f"Error `{e}' while finalizing experiment")
                 raise StateError('error', msg=str(e))
